Migrate to Next.js 16 and React 19 on Turbopack
Upgrades Next.js 14.2.35 -> 16.2.9 and React 18 -> 19.2.7, resolving the
RSC denial-of-service advisory CVE-2026-23869 (fixed in React 19.2.5+ and
backported into Next 16.2.9). Supersedes the failing Dependabot PR #107.

Why Turbopack: Next.js 16's webpack build path has a regression that
misclassifies every @next/mdx page as a Client Component, breaking the
`export const metadata` in all .mdx pages (reproduced on both 16.2.9 and
16.3 canary). Turbopack is the only working Next 16 path, so the two
webpack-coupled systems in this template were reworked for it.

Changes:
- Bump next, @next/mdx, eslint-config-next to 16.2.9; react/react-dom to
  19.2.7; @types/react(/dom) to 19; eslint to 9; framer-motion to 12,
  next-themes to 0.4, react-highlight-words to 0.21 (React 19 peers).
- MDX plugins: Turbopack can't pass plugin functions across the Rust/JS
  boundary, so the inline remark/rehype/recma plugins now live in
  src/mdx/plugins/* and are referenced by absolute module path from
  next.config.mjs (replacing src/mdx/{recma,rehype,remark}.mjs).
- Search index: the custom webpack loader in search.mjs (which Turbopack
  can't run) is split into a build-time Node generator (search-data.mjs,
  invoked from next.config.mjs -> .search-data.json) and a browser module
  (search.mjs) that builds the FlexSearch index from that JSON.
- Migrate `next lint` -> ESLint CLI flat config (eslint.config.mjs); the
  new React-Compiler react-hooks rules introduced by eslint-config-next 16
  are downgraded to warnings to avoid refactoring working template code.
- Fix React 19 type changes (isValidElement<T>, useRef initial arg,
  RefObject<T | null>) in Code/Sectionizer/SectionProvider.
- Remove stale pnpm-lock.yaml and the now-unused simple-functional-loader;
  ignore the generated search data.
