+++ package.json
@@ -6,7 +6,7 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "eslint ."
   },
   "browserslist": "defaults, not ie <= 11",
   "dependencies": {
@@ -15,7 +15,7 @@
     "@headlessui/tailwindcss": "^0.2.0",
     "@mdx-js/loader": "^3.0.0",
     "@mdx-js/react": "^3.0.0",
-    "@next/mdx": "16.2.9",
+    "@next/mdx": "^16.2.9",
     "@radix-ui/react-collapsible": "^1.1.12",
     "@sindresorhus/slugify": "^2.1.1",
     "@tailwindcss/typography": "^0.5.10",
@@ -32,7 +32,7 @@
     "framer-motion": "^12.23.24",
     "mdast-util-to-string": "^4.0.0",
     "mdx-annotations": "^0.1.1",
-    "next": "16.2.9",
+    "next": "^16.2.9",
     "next-themes": "^0.4.6",
     "react": "19.2.7",
     "react-dom": "19.2.7",
@@ -42,7 +42,6 @@
     "remark-html": "^16.0.1",
     "remark-mdx": "^3.0.0",
     "shiki": "^0.14.7",
-    "simple-functional-loader": "^1.2.1",
     "tailwind-merge": "^3.3.1",
     "tailwindcss": "^3.4.1",
     "typescript": "^5.3.3",
@@ -52,7 +51,7 @@
   },
   "devDependencies": {
     "eslint": "^9.39.4",
-    "eslint-config-next": "16.2.9",
+    "eslint-config-next": "^16.2.9",
     "prettier": "^3.3.2",
     "prettier-plugin-tailwindcss": "^0.6.5",
     "sharp": "0.33.1"
