+++ src/components/SectionProvider.tsx
@@ -16,7 +16,7 @@ export interface Section {
   title: string
   offsetRem?: number
   tag?: string
-  headingRef?: React.RefObject<HTMLHeadingElement>
+  headingRef?: React.RefObject<HTMLHeadingElement | null>
 }
 
 interface SectionState {
@@ -29,7 +29,7 @@ interface SectionState {
     offsetRem,
   }: {
     id: string
-    ref: React.RefObject<HTMLHeadingElement>
+    ref: React.RefObject<HTMLHeadingElement | null>
     offsetRem: number
   }) => void
 }
