--- src/mdx/recma.mjs
@@ -1,3 +1,0 @@
-import { mdxAnnotations } from 'mdx-annotations'
-
-export const recmaPlugins = [mdxAnnotations.recma]
