--- src/mdx/remark.mjs
@@ -1,4 +1,0 @@
-import { mdxAnnotations } from 'mdx-annotations'
-import remarkGfm from 'remark-gfm'
-
-export const remarkPlugins = [mdxAnnotations.remark, remarkGfm]
